a lot of changes
yeah.

for some reason i wasnt able to commit changes on the school computer today so i had to manually do this at home :D

basically, just made both the files and the core structure of (most of) the required game elements. i also made an (untested and terrible) code block/idea for adding characters, but this should probably be put in game.java.

- ⑨