workworkwork
a lot of changes. big rep to sam for making a brand new constructor method from scratch, as well as some points to anthony for starting reworks on the wishlist and the wished characters system [ not shown here bc they arent shared yet ] .

i reworked and finished the seriessearch method (again untested), and i added onto the methods in game. we're almost done, we got this.

- ⑨