sorta got the search method to work + cleaning
after maybe 30 mins of frustratingly staring at my computer screen, i got the search method in Game to work.

i also cleaned up some stuff in wishChar and in Character.

honestly, the project is due tomorrow, so this might be the last commit in a while. idk if we really are going to get this finished.

see you whenever.

- ⑨